+++ clang/bindings/python/clang/cindex.py
@@ -149,16 +149,14 @@ def b(x: str | bytes) -> bytes:
 # this by marshalling object arguments as void**.
 c_object_p: TType[_Pointer[Any]] = POINTER(c_void_p)
 
-
 ### Exception Classes ###
+
 
 class TranslationUnitLoadError(Exception):
     """Represents an error that occurred when loading a TranslationUnit.
 
     This is raised in the case where a TranslationUnit could not be
     instantiated due to failure in the libclang library.
-
-    FIXME: Make libclang expose additional error information in this scenario.
     """
 
     # A generic error code, no further details are available.
@@ -177,19 +175,18 @@ class TranslationUnitLoadError(Exception):
     # An AST deserialization error has occurred.
     ERROR_AST_READ_ERROR = 4
 
-    def __init__(self, enumeration: int | None, message: str):
-        if enumeration is not None:
-            assert isinstance(enumeration, int)
+    def __init__(self, enumeration: int, message: str):
+        assert isinstance(enumeration, int)
 
-            if enumeration < 1 or enumeration > 4:
-                raise Exception(
-                    "Encountered undefined CXError "
-                    "constant: %d. Please file a bug to have this "
-                    "value supported." % enumeration
-                )
+        if enumeration < 1 or enumeration > 4:
+            raise Exception(
+                "Encountered undefined CXError "
+                "constant: %d. Please file a bug to have this "
+                "value supported." % enumeration
+            )
 
         self.error_code = enumeration
-        Exception.__init__(self, "Error %d: %s" % (enumeration or 0, message))
+        Exception.__init__(self, "Error %d: %s" % (enumeration, message))
 
 
 class TranslationUnitSaveError(Exception):
@@ -3111,7 +3108,8 @@ class TranslationUnit(ClangObject):
 
         unsaved_array = cls.process_unsaved_files(unsaved_files)
 
-        ptr = conf.lib.clang_parseTranslationUnit(
+        ptr = c_object_p()
+        errc = conf.lib.clang_parseTranslationUnit2(
             index,
             os.fspath(filename) if filename is not None else None,
             args_array,
@@ -3119,11 +3117,11 @@ class TranslationUnit(ClangObject):
             unsaved_array,
             len(unsaved_files),
             options,
+            byref(ptr),
         )
 
-        if not ptr:
-            # FIXME: use clang_parseTranslationUnit2 to preserve error code
-            raise TranslationUnitLoadError(None, "Error parsing translation unit.")
+        if errc != 0:
+            raise TranslationUnitLoadError(errc, "Error parsing translation unit.")
 
         return cls(ptr, index=index)
 
@@ -3145,10 +3143,15 @@ class TranslationUnit(ClangObject):
         if index is None:
             index = Index.create()
 
-        ptr = conf.lib.clang_createTranslationUnit(index, os.fspath(filename))
-        if not ptr:
-            # FIXME: use clang_createTranslationUnit2 to preserve error code
-            raise TranslationUnitLoadError(None, filename)
+        ptr = c_object_p()
+        errc = conf.lib.clang_createTranslationUnit2(
+            index,
+            os.fspath(filename),
+            byref(ptr)
+        )
+
+        if errc != 0:
+            raise TranslationUnitLoadError(errc, filename)
 
         return cls(ptr=ptr, index=index)
 
@@ -3293,11 +3296,11 @@ class TranslationUnit(ClangObject):
             unsaved_files = []
 
         unsaved_files_array = self.process_unsaved_files(unsaved_files)
-        result = conf.lib.clang_reparseTranslationUnit(
+        errc = conf.lib.clang_reparseTranslationUnit(
             self, len(unsaved_files), unsaved_files_array, options
         )
-        if result != 0:
-            raise TranslationUnitLoadError(result, 'Error reparsing TranslationUnit.')
+        if errc != 0:
+            raise TranslationUnitLoadError(errc, 'Error reparsing TranslationUnit.')
 
     def save(self, filename):
         """Saves the TranslationUnit to a file.
@@ -3751,6 +3754,7 @@ functionList: list[LibFunc] = [
     ("clang_codeCompleteGetNumDiagnostics", [CodeCompletionResults], c_int),
     ("clang_createIndex", [c_int, c_int], c_object_p),
     ("clang_createTranslationUnit", [Index, c_interop_string], c_object_p),
+    ("clang_createTranslationUnit2", [Index, c_interop_string, POINTER(c_object_p)], c_int),
     ("clang_CXRewriter_create", [TranslationUnit], c_object_p),
     ("clang_CXRewriter_dispose", [Rewriter]),
     ("clang_CXRewriter_insertTextBefore", [Rewriter, SourceLocation, c_interop_string]),
@@ -3944,6 +3948,11 @@ functionList: list[LibFunc] = [
         "clang_parseTranslationUnit",
         [Index, c_interop_string, c_void_p, c_int, c_void_p, c_int, c_int],
         c_object_p,
+    ),
+    (
+        "clang_parseTranslationUnit2",
+        [Index, c_interop_string, c_void_p, c_int, c_void_p, c_int, c_int, POINTER(c_object_p)],
+        c_int,
     ),
     ("clang_reparseTranslationUnit", [TranslationUnit, c_int, c_void_p, c_int], c_int),
     ("clang_saveTranslationUnit", [TranslationUnit, c_interop_string, c_uint], c_int),